requirements.txt
altgraph==0.17.2
click==8.1.3
colorama==0.4.4
decorator==5.1.1
fire==0.4.0
Flask==2.1.2
future==0.18.2
importlib-metadata==4.11.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
llvmlite==0.37.0
lsnms==0.1.2
MarkupSafe==2.1.1
networkx==2.5
numba==0.54.1
numpy==1.19.5
opencv-python==4.5.5.64
openvino==2021.4.2
pefile==2022.5.30
Pillow==9.1.1
pyaml==21.10.1
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pywin32-ctypes==0.2.0
PyYAML==6.0
sahi==0.8.8
scikit-learn==1.0.2
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tqdm==4.64.0
Werkzeug==2.1.2
zipp==3.8.0
